pyproject.toml
[tool.poetry]
name = "llmreflect"
version = "0.1.2"
description = "a package for llm self-reflection"
authors = ["alchemistwu <alchemistwu0521@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.3"
python-decouple = "^3.8"
openai = "^0.27.8"
psycopg2-binary = "^2.9.6"
langchain = "^0.0.219"
SQLAlchemy = "^2.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
